Menu added
I have defined a way to go back to the main menu every time the user
click on the “classified exchange app”. Once click on it, it will
return to the main menu. Nonetheless, when the user shrink the window,
the toolbar will show a small icon that will list the content. The idea
behind that is we have a workable menu.

todo: add more content to it