First commit for this new feature.
Added a new menu item 'Administrator' with a sub-item 'Categories' that
will allow an admin to add a category to an ad.

The category menu will list all the categories available and will allow
the administrator to add new and delete categories.

The category contains a name and a description.