Fix password validation
There was a bug with the password progress bar. I've changed some functions to keep track of the validation. Essentially 4 regex patterns were added to verify how secure the user's password is. The progress bar advances as the user meets all the requirements of security that we've shown on screen.